requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pinecone==5.4.2
langchain~=0.3.25
langchain-openai==0.3.9
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.7.4
pydantic-settings~=2.9.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
cryptography==41.0.5 
langchain-community~=0.3.25
langchain-core~=0.3.65
streamlit==1.32.0
requests==2.31.0
duckduckgo_search~=8.0.3